requirements.txt
streamlit
python-dotenv
requests
google-generativeai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
schedule
moviepy
edge-tts
pandas
numpy
